requirements.txt
python-dotenv==1.0.1
python-slugify==8.0.4